move resulting_type_kind to lowering, don't handle functions
